requirements 2.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.7
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
pytest==7.4.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
